+++ bin/example.php
@@ -5,7 +5,7 @@ if (!defined('DS')) {
     define('DS', DIRECTORY_SEPARATOR);
 }
 
-declare(ticks = 100);
+declare(ticks = 1000);
 
 require_once(__DIR__ . DS . '..' . DS . 'examples' . DS . 'bootstrap.php');
 
@@ -15,6 +15,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\ArrayInput;
 
 use HRDNS\Examples\Socket\EchoServer;
+use HRDNS\Examples\Socket\WebSocketServer;
 
 class Application extends BaseApplication
 {
@@ -33,6 +34,7 @@ class Application extends BaseApplication
     {
         $commands = parent::getDefaultCommands();
         $commands[] = new EchoServer();
+        $commands[] = new WebSocketServer();
         return $commands;
     }
 
@@ -53,7 +55,7 @@ class Application extends BaseApplication
         }
         if (!$input->getFirstArgument()) {
             $input = new ArrayInput(
-                array ('list')
+                ['list']
             );
         }
         return parent::doRun($input, $output);